+++ src/onegov.swissvotes/onegov/swissvotes/tests/test_views_vote.py
@@ -285,3 +285,62 @@ def test_vote_upload(swissvotes_app, attachments):
         assert page.content_type == 'application/pdf'
         assert page.content_length
         assert page.body
+
+
+def test_vote_pagination(swissvotes_app):
+    for day, number in ((1, '100'), (2, '101.1'), (2, '101.2'), (3, '102')):
+        swissvotes_app.session().add(
+            SwissVote(
+                bfs_number=Decimal(number),
+                date=date(1990, 6, day),
+                decade=NumericRange(1990, 1999),
+                legislation_number=4,
+                legislation_decade=NumericRange(1990, 1994),
+                title="Vote",
+                keyword="Keyword",
+                votes_on_same_day=2,
+                _legal_form=3,
+                initiator="Initiator",
+            )
+        )
+    commit()
+
+    client = Client(swissvotes_app)
+    client.get('/locale/de_CH').follow()
+
+    # 102
+    page = client.get('/').maybe_follow().click("Abstimmungen")
+    page = page.click("Details", index=0)
+    assert "<td>102</td>" in page
+    assert "Vorherige Abstimmung" in page
+    assert "Nächste Abstimmung" not in page
+
+    # 101.2
+    page = page.click("Vorherige Abstimmung")
+    assert "<td>101.2</td>" in page
+    assert "Vorherige Abstimmung" in page
+    assert "Nächste Abstimmung" in page
+
+    # 101.1
+    page = page.click("Vorherige Abstimmung")
+    assert "<td>101.1</td>" in page
+    assert "Vorherige Abstimmung" in page
+    assert "Nächste Abstimmung" in page
+
+    # 100
+    page = page.click("Vorherige Abstimmung")
+    assert "<td>100</td>" in page
+    assert "Vorherige Abstimmung" not in page
+    assert "Nächste Abstimmung" in page
+
+    # 101.1
+    page = page.click("Nächste Abstimmung")
+    assert "<td>101.1</td>" in page
+
+    # 101.2
+    page = page.click("Nächste Abstimmung")
+    assert "<td>101.2</td>" in page
+
+    # 102
+    page = page.click("Nächste Abstimmung")
+    assert "<td>102</td>" in page