+++ src/onegov.swissvotes/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds a pagination to the vote details view.
+  [msom]
+
 0.0.26 (2018-12-18)
 ~~~~~~~~~~~~~~~~~~~